+++ src/components/ids-data-grid/ids-data-grid.ts
@@ -45,6 +45,10 @@ import IdsDataGridCell from './ids-data-grid-cell';
 import { ExcelColumn } from '../../utils/ids-excel-exporter/ids-worksheet-templates';
 import IdsLoadingIndicator from '../ids-loading-indicator/ids-loading-indicator';
 
+// Types
+import type IdsHyperlink from '../ids-hyperlink/ids-hyperlink';
+import type IdsButton from '../ids-button/ids-button';
+
 const Base = IdsPagerMixin(
   IdsDataGridSaveSettingsMixin(
     IdsDataGridTooltipMixin(
@@ -638,11 +642,31 @@ export default class IdsDataGrid extends Base {
       const isClickable = isButton || isHyperlink;
       const column: IdsDataGridColumn = this.visibleColumns[cellNum];
 
-      // Focus Cell
+      // Focus cell
       this.setActiveCell(cellNum, rowNum, isHyperlink);
       // Handle click callbacks
       if (isClickable && column.click !== undefined && !e.target?.getAttribute('disabled')) {
         column.click(this.data[rowNum], this.visibleColumns[cellNum], e);
+      }
+
+      // hide menus on body click
+      this.hideOpenMenus();
+
+      // Handle popup menus
+      if (isClickable && column?.menuId !== undefined && !e.target?.getAttribute('disabled')) {
+        const menuEl = document.querySelector<IdsPopupMenu>(`#${column.menuId}`);
+        if (!menuEl) return;
+        menuEl.target = null;
+        menuEl.hide();
+        menuEl.triggerType = 'immediate-closed';
+        menuEl.target = e.target;
+        menuEl.align = 'bottom, left';
+        menuEl.arrow = 'bottom';
+        menuEl.triggerType = 'immediate';
+        menuEl.offEvent('selected', menuEl);
+        menuEl.onEvent('selected', menuEl, (evt: CustomEvent) => {
+          column?.selected?.(this.data[rowNum], this.visibleColumns[cellNum], evt);
+        });
       }
 
       // Fires for each row that is clicked
@@ -798,6 +822,10 @@ export default class IdsDataGrid extends Base {
       const activeRowIndex = nextCell?.rowIndex || 0;
       nextCell?.focus?.();
 
+      if (nextCell && nextCell.columnIndex >= 0) {
+        this.setActiveCell(nextCell.columnIndex, activeRowIndex, false);
+      }
+
       // Handle row selection
       const movingVertical = key === 'ArrowDown' || key === 'ArrowUp';
       if ((this.rowSelection === 'mixed' || this.rowSelection === 'multiple') && movingVertical && e.shiftKey) {
@@ -856,10 +884,21 @@ export default class IdsDataGrid extends Base {
 
     // Follow links with keyboard and start editing
     this.listen(['Enter'], this, (e: KeyboardEvent) => {
-      // if (this.openMenu) return;
-      // if (!this.activeCellCanClose()) return false;
-
       const cellNode = this.cellLastActive;
+
+      // Hyperlink
+      if (cellNode?.classList.contains('formatter-hyperlink')) {
+        cellNode.querySelector<IdsHyperlink>('ids-hyperlink')?.container?.click();
+        return;
+      }
+
+      // Button
+      if (cellNode?.classList.contains('formatter-button')) {
+        cellNode.querySelector<IdsButton>('ids-button')?.button?.click();
+        return;
+      }
+
+      // Editahe
       if (!cellNode?.isEditing) {
         cellNode?.startCellEdit();
         return;
@@ -893,8 +932,6 @@ export default class IdsDataGrid extends Base {
 
     // Edit Next
     this.listen(['Tab'], this, (e: KeyboardEvent) => {
-      // if (this.openMenu) return;
-      // if (!this.activeCellCanClose()) return false;
       if (!this.isEditable) {
         return;
       }
@@ -3192,6 +3229,17 @@ export default class IdsDataGrid extends Base {
     if (this.activeCellEditor?.column?.editor?.inline) return false;
 
     return true;
+  }
+
+  /**
+   * Hide any open context menus
+   * @private
+   */
+  hideOpenMenus() {
+    const openMenus = document.querySelector<IdsPopupMenu>('ids-popup-menu:not([hidden])');
+    if (openMenus) {
+      openMenus.hide();
+    }
   }
 
   /**