+++ tests/ids-input/ids-input-validation.spec.ts
@@ -29,7 +29,7 @@ test.describe('IdsInput validation tests', () => {
       elem.checkValidation();
     });
     expect(await input.evaluate((elem: IdsInput) => elem.input?.getAttribute('aria-invalid'))).toEqual('true');
-    expect(await input.evaluate((elem: IdsInput) => elem.input?.getAttribute('aria-describedby'))).toEqual('test-input-error');
+    expect(await input.evaluate((elem: IdsInput) => elem.input?.getAttribute('aria-describedby'))).toEqual('test-input-internal-error');
     const results = await input.evaluate((elem: IdsInput) => {
       const msgEl = elem.container?.querySelector('.validation-message');
 
@@ -43,7 +43,7 @@ test.describe('IdsInput validation tests', () => {
 
     expect(results.msgEl).toBeDefined();
     expect(results.validationId).toEqual('required');
-    expect(results.id).toEqual('test-input-error');
+    expect(results.id).toEqual('test-input-internal-error');
     expect(results.count).toEqual(1);
 
     await input.evaluate((elem: IdsInput) => {
@@ -256,7 +256,7 @@ test.describe('IdsInput validation tests', () => {
 
     expect(results.msgEl).toBeDefined();
     expect(results.validationId).toEqual('email');
-    expect(results.id).toEqual('test-input-error');
+    expect(results.id).toEqual('test-input-internal-error');
     expect(results.count).toEqual(1);
 
     await input.evaluate((elem: IdsInput) => {
@@ -333,7 +333,7 @@ test.describe('IdsInput validation tests', () => {
     let results = await getResults();
     expect(results.msgEl).toBeDefined();
     expect(results.validationId).toEqual('error');
-    expect(results.id).toEqual('test-input-error');
+    expect(results.id).toEqual('test-input-internal-error');
     expect(results.count).toEqual(1);
     expect(results.textContent).toEqual('Error test');
     // should add input alert message
@@ -344,7 +344,7 @@ test.describe('IdsInput validation tests', () => {
     results = await getResults();
     expect(results.msgEl).toBeDefined();
     expect(results.validationId).toEqual('alert');
-    expect(results.id).toEqual('test-input-alert');
+    expect(results.id).toEqual('test-input-internal-alert');
     expect(results.count).toEqual(1);
     expect(results.textContent).toEqual('Alert test');
     // should add input success message
@@ -355,7 +355,7 @@ test.describe('IdsInput validation tests', () => {
     results = await getResults();
     expect(results.msgEl).toBeDefined();
     expect(results.validationId).toEqual('success');
-    expect(results.id).toEqual('test-input-success');
+    expect(results.id).toEqual('test-input-internal-success');
     expect(results.count).toEqual(1);
     expect(results.textContent).toEqual('Success test');
     // should add input info message
@@ -366,7 +366,7 @@ test.describe('IdsInput validation tests', () => {
     results = await getResults();
     expect(results.msgEl).toBeDefined();
     expect(results.validationId).toEqual('info');
-    expect(results.id).toEqual('test-input-info');
+    expect(results.id).toEqual('test-input-internal-info');
     expect(results.count).toEqual(1);
     expect(results.textContent).toEqual('Info test');
     // should add input default icon message
@@ -377,7 +377,7 @@ test.describe('IdsInput validation tests', () => {
     results = await getResults();
     expect(results.msgEl).toBeDefined();
     expect(results.validationId).toEqual('icon');
-    expect(results.id).toEqual('test-input-icon');
+    expect(results.id).toEqual('test-input-internal-icon');
     expect(results.count).toEqual(1);
     expect(results.textContent).toEqual('test');
     // should add input custom icon message
@@ -390,7 +390,7 @@ test.describe('IdsInput validation tests', () => {
     results = await getResults();
     expect(results.msgEl).toBeDefined();
     expect(results.validationId).toEqual('icon-custom');
-    expect(results.id).toEqual('test-input-icon');
+    expect(results.id).toEqual('test-input-internal-icon');
     expect(results.count).toEqual(1);
     expect(results.textContent).toEqual('test');
   });